Rename Store to State.
- Fix a bug where nickname updating didn't actually work for
  channelUsers section of the database.
- Add more flexibility in the server's ability to lock up the State.
  Using OpenState/CloseState or UsingState.
- Rename Server's state variable to status.

+++ data/user_modes.go
@@ -2,7 +2,7 @@ package data
 
 // UserModes provides basic modes for channels and users.
 type UserModes struct {
-	modes int
+	modes byte
 	*UserModeKinds
 }
